requirements.txt
fastapi==0.95.2
sqlalchemy==1.4.46
alembic==1.10.4
psycopg2-binary==2.9.6
uvicorn==0.22.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
redis==4.6.0
aiofiles==23.1.0
Pillow==9.5.0
pika==1.3.2
pydantic==1.10.12
python-dotenv==1.0.0
python-multipart==0.0.6
orjson